+++ datacube/index/abstract.py
@@ -825,7 +825,8 @@ class AbstractDatasetResource(ABC):
 
     @abstractmethod
     def add(self, dataset: Dataset,
-            with_lineage: bool = True
+            with_lineage: bool = True,
+            archive_less_mature: bool = False,
            ) -> Dataset:
         """
         Add ``dataset`` to the index. No-op if it is already present.
@@ -836,6 +837,10 @@ class AbstractDatasetResource(ABC):
            - ``True (default)`` attempt adding lineage datasets if missing
            - ``False`` record lineage relations, but do not attempt
              adding lineage datasets to the db
+
+        :param archive_less_mature:
+            - ``True`` search for less mature versions of the dataset
+            and archive them
 
         :return: Persisted Dataset model
         """
@@ -874,12 +879,14 @@ class AbstractDatasetResource(ABC):
     @abstractmethod
     def update(self,
                dataset: Dataset,
-               updates_allowed: Optional[Mapping[Offset, AllowPolicy]] = None
+               updates_allowed: Optional[Mapping[Offset, AllowPolicy]] = None,
+               archive_less_mature: bool = False,
               ) -> Dataset:
         """
         Update dataset metadata and location
         :param Dataset dataset: Dataset model with unpersisted updates
         :param updates_allowed: Allowed updates
+        :param archive_less_mature: Find and archive less mature datasets
         :return: Persisted dataset model
         """
 
@@ -890,6 +897,37 @@ class AbstractDatasetResource(ABC):
 
         :param Iterable[Union[str,UUID]] ids: list of dataset ids to archive
         """
+
+    def archive_less_mature(self, ds: Dataset) -> None:
+        """
+        Archive less mature versions of a dataset
+
+        :param Dataset ds: dataset to search
+        """
+        less_mature = []
+        dupes = self.search(product=ds.product.name,
+                            region_code=ds.metadata.region_code,
+                            time=ds.metadata.time)
+        for dupe in dupes:
+            if dupe.id == ds.id:
+                continue
+            if dupe.metadata.dataset_maturity == ds.metadata.dataset_maturity:
+                # Duplicate has the same maturity, which one should be archived is unclear
+                raise ValueError(
+                    f"A dataset with the same maturity as dataset {ds.id} already exists, "
+                    f"with id: {dupe.id}"
+                )
+            if dupe.metadata.dataset_maturity < ds.metadata.dataset_maturity:
+                # Duplicate is more mature than dataset
+                # Note that "final" < "nrt"
+                raise ValueError(
+                    f"A more mature version of dataset {ds.id} already exists, with id: "
+                    f"{dupe.id} and maturity: {dupe.metadata.dataset_maturity}"
+                )
+            less_mature.append(dupe.id)
+        self.archive(less_mature)
+        for lm_ds in less_mature:
+            _LOG.info(f"Archived less mature dataset: {lm_ds}")
 
     @abstractmethod
     def restore(self, ids: Iterable[DSID]) -> None: