+++ datacube/index/postgis/_datasets.py
@@ -137,7 +137,7 @@ class DatasetResource(AbstractDatasetResource, IndexResourceAddIn):
                 map((lambda x: UUID(x) if isinstance(x, str) else x), ids_)]
 
     def add(self, dataset: Dataset,
-            with_lineage: bool = True) -> Dataset:
+            with_lineage: bool = True, archive_less_mature: bool = False) -> Dataset:
         """
         Add ``dataset`` to the index. No-op if it is already present.
 
@@ -147,6 +147,10 @@ class DatasetResource(AbstractDatasetResource, IndexResourceAddIn):
            - ``True (default)`` attempt adding lineage datasets if missing
            - ``False`` record lineage relations, but do not attempt
              adding lineage datasets to the db
+
+        :param archive_less_mature:
+            - ``True`` search for less mature versions of the dataset
+            and archive them
 
         :rtype: Dataset
         """
@@ -169,6 +173,8 @@ class DatasetResource(AbstractDatasetResource, IndexResourceAddIn):
                 # 1c. Store locations
                 if dataset.uris is not None:
                     self._ensure_new_locations(dataset, transaction=transaction)
+            if archive_less_mature:
+                self.archive_less_mature(dataset)
 
         return dataset
 
@@ -305,11 +311,12 @@ class DatasetResource(AbstractDatasetResource, IndexResourceAddIn):
 
         return not bad_changes, good_changes, bad_changes
 
-    def update(self, dataset: Dataset, updates_allowed=None):
+    def update(self, dataset: Dataset, updates_allowed=None, archive_less_mature=False):
         """
         Update dataset metadata and location
         :param Dataset dataset: Dataset to update
         :param updates_allowed: Allowed updates
+        :param archive_less_mature: Find and archive less mature datasets
         :rtype: Dataset
         """
         existing = self.get(dataset.id)
@@ -342,6 +349,8 @@ class DatasetResource(AbstractDatasetResource, IndexResourceAddIn):
                 raise ValueError("Failed to update dataset %s..." % dataset.id)
             transaction.update_spindex(dsids=[dataset.id])
             transaction.update_search_index(dsids=[dataset.id])
+            if archive_less_mature:
+                self._archive_less_mature(dataset)
 
         self._ensure_new_locations(dataset, existing)