+++ integration_tests/conftest.py
@@ -197,6 +197,24 @@ def africa_s2_product_doc():
     return get_eo3_test_data_doc("s2_africa_product.yaml")
 
 
+@pytest.fixture
+def final_dataset_doc():
+    return (
+        get_eo3_test_data_doc("final_dataset.yaml"),
+        's3://dea-public-data/baseline/ga_ls8c_ard_3/090/086/2023/04/30'
+        'ga_ls8c_ard_3-2-1_090086_2023-04-30_final.stac-item.json'
+    )
+
+
+@pytest.fixture
+def nrt_dataset_doc():
+    return (
+        get_eo3_test_data_doc("nrt_dataset.yaml"),
+        's3://dea-public-data/baseline/ga_ls8c_ard_3/090/086/2023/04/30_nrt'
+        'ga_ls8c_ard_3-2-1_090086_2023-04-30_nrt.stac-item.json'
+    )
+
+
 def doc_to_ds(index, product_name, ds_doc, ds_path):
     from datacube.index.hl import Doc2Dataset
     resolver = Doc2Dataset(index, products=[product_name], verify_lineage=False)
@@ -279,6 +297,28 @@ def africa_eo3_dataset(index, africa_s2_eo3_product, eo3_africa_dataset_doc):
     return doc_to_ds(index,
                      africa_s2_eo3_product.name,
                      *eo3_africa_dataset_doc)
+
+
+@pytest.fixture
+def nrt_dataset(index, extended_eo3_metadata_type, ls8_eo3_product, nrt_dataset_doc):
+    ds_doc = nrt_dataset_doc[0]
+    ds_path = nrt_dataset_doc[1]
+    from datacube.index.hl import Doc2Dataset
+    resolver = Doc2Dataset(index, products=[ls8_eo3_product.name], verify_lineage=False)
+    ds, err = resolver(ds_doc, ds_path)
+    assert err is None and ds is not None
+    return ds
+
+
+@pytest.fixture
+def final_dataset(index, extended_eo3_metadata_type, ls8_eo3_product, final_dataset_doc):
+    ds_doc = final_dataset_doc[0]
+    ds_path = final_dataset_doc[1]
+    from datacube.index.hl import Doc2Dataset
+    resolver = Doc2Dataset(index, products=[ls8_eo3_product.name], verify_lineage=False)
+    ds, err = resolver(ds_doc, ds_path)
+    assert err is None and ds is not None
+    return ds
 
 
 @pytest.fixture